requirements.txt
anthropic
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb
sentence-transformers
pdfplumber
python-dotenv
notion-client
supabase
streamlit